ternary to module
