+++ polars/polars-lazy/src/physical_plan/expressions/default.rs
@@ -8,28 +8,6 @@ use rayon::prelude::*;
 use std::borrow::Cow;
 use std::sync::Arc;
 
-pub struct TernaryExpr {
-    pub predicate: Arc<dyn PhysicalExpr>,
-    pub truthy: Arc<dyn PhysicalExpr>,
-    pub falsy: Arc<dyn PhysicalExpr>,
-    pub expr: Expr,
-}
-
-impl PhysicalExpr for TernaryExpr {
-    fn as_expression(&self) -> &Expr {
-        &self.expr
-    }
-    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
-        let mask_series = self.predicate.evaluate(df, state)?;
-        let mask = mask_series.bool()?;
-        let truthy = self.truthy.evaluate(df, state)?;
-        let falsy = self.falsy.evaluate(df, state)?;
-        truthy.zip_with(&mask, &falsy)
-    }
-    fn to_field(&self, input_schema: &Schema) -> Result<Field> {
-        self.truthy.to_field(input_schema)
-    }
-}
 
 pub struct ApplyExpr {
     pub input: Arc<dyn PhysicalExpr>,
